[ui] [installer] Update setup page to support wifi autoconfig
