+++ frontend/src/components/Wifi/WifiHostapd.js
@@ -1,6 +1,9 @@
 import React, { useContext, useEffect, useState } from 'react'
 
 import { wifiAPI } from 'api'
+
+import {generateCapabilitiesString, generateConfigForBand, sortConf, isSPRCompat} from 'api/Wifi'
+
 import { AlertContext } from 'AppContext'
 
 import { Platform } from 'react-native'
@@ -84,338 +87,6 @@ import { Select } from 'components/Select'
 
 import WifiChannelParameters from 'components/Wifi/WifiChannelParameters'
 import { ListHeader } from 'components/List'
-
-const default5Ghz = {
-  ap_isolate: 1,
-  auth_algs: 1,
-  channel: 36,
-  country_code: 'US',
-  ctrl_interface: '/state/wifi/control_wlan',
-  ht_capab:
-    '[LDPC][HT40+][HT40-][GF][SHORT-GI-20][SHORT-GI-40][TX-STBC][RX-STBC1]',
-  hw_mode: 'a',
-  ieee80211ac: 1,
-  ieee80211d: 1,
-  ieee80211n: 1,
-  ieee80211w: 1,
-  interface: 'wlan',
-  multicast_to_unicast: 1,
-  per_sta_vif: 1,
-  preamble: 1,
-  rsn_pairwise: 'CCMP',
-  sae_psk_file: '/configs/wifi/sae_passwords',
-  ssid: 'TestLab',
-  tdls_prohibit: 1,
-  vht_capab:
-    '[RXLDPC][SHORT-GI-80][TX-STBC-2BY1][RX-STBC-1][MAX-A-MPDU-LEN-EXP3][RX-ANTENNA-PATTERN][TX-ANTENNA-PATTERN]',
-  vht_oper_centr_freq_seg0_idx: 42,
-  vht_oper_chwidth: 1,
-  wmm_enabled: 1,
-  wpa: 2,
-  wpa_disable_eapol_key_retries: 1,
-  wpa_key_mgmt: 'WPA-PSK WPA-PSK-SHA256 SAE',
-  wpa_psk_file: '/configs/wifi/wpa2pskfile'
-}
-
-const default2Ghz = {
-  ap_isolate: 1,
-  auth_algs: 1,
-  channel: 1,
-  country_code: 'US',
-  ctrl_interface: '/state/wifi/control_wlan',
-  ht_capab:
-    '[LDPC][HT40+][HT40-][GF][SHORT-GI-20][SHORT-GI-40][TX-STBC][RX-STBC1]',
-  hw_mode: 'g',
-  ieee80211ac: 1,
-  ieee80211d: 1,
-  ieee80211n: 1,
-  ieee80211w: 1,
-  interface: 'wlan',
-  multicast_to_unicast: 1,
-  per_sta_vif: 1,
-  preamble: 1,
-  rsn_pairwise: 'CCMP',
-  sae_psk_file: '/configs/wifi/sae_passwords',
-  ssid: 'TestLab',
-  tdls_prohibit: 1,
-  wmm_enabled: 1,
-  wpa: 2,
-  wpa_disable_eapol_key_retries: 1,
-  wpa_key_mgmt: 'WPA-PSK WPA-PSK-SHA256 SAE',
-  wpa_psk_file: '/configs/wifi/wpa2pskfile'
-}
-
-const default6GHz = JSON.parse(JSON.stringify(default5Ghz))
-
-const htCapab = [
-  '[LDPC]',
-  '[HT40-]',
-  '[HT40+]',
-  '[GF]',
-  '[SHORT-GI-20]',
-  '[SHORT-GI-40]',
-  '[TX-STBC]',
-  '[RX-STBC1]',
-  '[RX-STBC12]',
-  '[RX-STBC123]',
-  '[DELAYED-BA]',
-  '[MAX-AMSDU-7935]',
-  '[DSSS_CCK-40]',
-  '[40-INTOLERANT]',
-  '[LSIG-TXOP-PROT]'
-]
-
-const vhtCapab = [
-  '[MAX-MPDU-7991]',
-  '[MAX-MPDU-11454]',
-  '[VHT160]',
-  '[VHT160-80PLUS80]',
-  '[RXLDPC]',
-  '[SHORT-GI-80]',
-  '[SHORT-GI-160]',
-  '[TX-STBC-2BY1]',
-  '[RX-STBC-1]',
-  '[RX-STBC-12]',
-  '[RX-STBC-123]',
-  '[RX-STBC-1234]',
-  '[SU-BEAMFORMER]',
-  '[SU-BEAMFORMEE]',
-  '[BF-ANTENNA-2]',
-  '[BF-ANTENNA-3]',
-  '[BF-ANTENNA-4]',
-  '[SOUNDING-DIMENSION-2]',
-  '[SOUNDING-DIMENSION-3]',
-  '[SOUNDING-DIMENSION-4]',
-  '[MU-BEAMFORMER]',
-  '[VHT-TXOP-PS]',
-  '[HTC-VHT]',
-  '[MAX-A-MPDU-LEN-EXP7]',
-  '[MAX-A-MPDU-LEN-EXP6]',
-  '[MAX-A-MPDU-LEN-EXP5]',
-  '[MAX-A-MPDU-LEN-EXP4]',
-  '[MAX-A-MPDU-LEN-EXP3]',
-  '[MAX-A-MPDU-LEN-EXP2]',
-  '[MAX-A-MPDU-LEN-EXP1]',
-  '[VHT-LINK-ADAPT2]',
-  '[VHT-LINK-ADAPT3]',
-  '[RX-ANTENNA-PATTERN]',
-  '[TX-ANTENNA-PATTERN]'
-]
-
-function mapHTCapabilities(capStr) {
-  // Extract HT capabilities bitfield from the
-  const htCapa = parseInt(capStr, 16)
-
-  const parsedOutput = []
-
-  // Check capabilities using bitwise operations
-  if (htCapa & (1 << 0)) {
-    parsedOutput.push('[LDPC]')
-  }
-
-  if (htCapa & (1 << 1)) {
-    //supports HT40+/-
-    parsedOutput.push('[HT40+]')
-    parsedOutput.push('[HT40-]')
-  }
-
-  if (htCapa & (1 << 4)) {
-    parsedOutput.push('[GF]')
-  }
-  if (htCapa & (1 << 5)) {
-    parsedOutput.push('[SHORT-GI-20]')
-  }
-  if (htCapa & (1 << 6)) {
-    parsedOutput.push('[SHORT-GI-40]')
-  }
-  if (htCapa & (1 << 7)) {
-    parsedOutput.push('[TX-STBC]')
-  }
-
-  if (htCapa & (1 << 8) && htCapa & (1 << 9)) {
-    parsedOutput.push('[RX-STBC123]')
-  } else if (htCapa & (1 << 9)) {
-    parsedOutput.push('[RX-STBC12]')
-  } else if (htCapa & (1 << 8)) {
-    parsedOutput.push('[RX-STBC1]')
-  }
-
-  if (htCapa & (1 << 10)) {
-    parsedOutput.push('[DELAYED-BA]')
-  }
-  if (htCapa & (1 << 11)) {
-    parsedOutput.push('[MAX-AMSDU-7935]')
-  }
-  if (htCapa & (1 << 12)) {
-    parsedOutput.push('[DSSS_CCK-40]')
-  }
-  if (htCapa & (1 << 14)) {
-    parsedOutput.push('[40-INTOLERANT]')
-  }
-
-  if (htCapa & (1 << 15)) {
-    parsedOutput.push('[LSIG-TXOP-PROT]')
-  }
-
-  return parsedOutput
-}
-
-function mapVHTCapabilities(capStr) {
-  const capa = parseInt(capStr, 16)
-
-  const parsedOutput = []
-
-  // Check capabilities using bitwise operations
-  if (capa & (1 << 0)) {
-    parsedOutput.push('[MAX-MPDU-7991]')
-  }
-  if (capa & (1 << 1)) {
-    parsedOutput.push('[MAX-MPDU-11454]')
-  }
-  if (capa & (1 << 2)) {
-    parsedOutput.push('[VHT160]')
-  }
-  //note: if both 2 and 3 are set, then this is incorrect
-  if (capa & (1 << 3)) {
-    parsedOutput.push('[VHT160-80PLUS80]')
-  }
-  if (capa & (1 << 4)) {
-    parsedOutput.push('[RXLDPC]')
-  }
-  if (capa & (1 << 5)) {
-    parsedOutput.push('[SHORT-GI-80]')
-  }
-  if (capa & (1 << 6)) {
-    parsedOutput.push('[SHORT-GI-160]')
-  }
-  if (capa & (1 << 7)) {
-    parsedOutput.push('[TX-STBC-2BY1]')
-  }
-
-  switch ((capa & 0x700) >> 8) {
-    case 1:
-      parsedOutput.push('[RX-STBC-1]')
-      break
-    case 2:
-      parsedOutput.push('[RX-STBC-12]')
-      break
-    case 3:
-      parsedOutput.push('[RX-STBC-123]')
-      break
-    case 4:
-      parsedOutput.push('[RX-STBC-1234]')
-      break
-    case 5:
-      parsedOutput.push('[RX-STBC-1234]')
-      break //hostapd does not support 5 yet
-    case 6:
-      parsedOutput.push('[RX-STBC-1234]')
-      break //hostapd does not support 6 yet
-    case 7:
-      parsedOutput.push('[RX-STBC-1234]')
-      break //hostapd does not support 7 yet
-    default:
-      break
-  }
-
-  if (capa & (1 << 11)) {
-    parsedOutput.push('[SU-BEAMFORMER]')
-  }
-  if (capa & (1 << 12)) {
-    parsedOutput.push('[SU-BEAMFORMEE]')
-  }
-
-  let beamformingAntennas = ((capa & 0xe000) >> 13) + 1
-  if (beamformingAntennas > 1) {
-    parsedOutput.push('[BF-ANTENNA-' + beamformingAntennas + ']')
-  }
-
-  let soundingDimensions = ((capa & 0x70000) >> 16) + 1
-  if (soundingDimensions > 1) {
-    parsedOutput.push('[SOUNDING-DIMENSION-' + soundingDimensions + ']')
-  }
-
-  if (capa & (1 << 19)) {
-    parsedOutput.push('[MU-BEAMFORMER]')
-  }
-  if (capa & (1 << 21)) {
-    parsedOutput.push('[VHT-TXOP-PS]')
-  }
-  if (capa & (1 << 22)) {
-    parsedOutput.push('[HTC-VHT]')
-  }
-
-  let aMPDUExp = (capa & 0x3800000) >> 23
-
-  if (aMPDUExp > 0) {
-    parsedOutput.push('[MAX-A-MPDU-LEN-EXP' + aMPDUExp + ']')
-  }
-
-  if (capa & (1 << 28)) {
-    parsedOutput.push('[RX-ANTENNA-PATTERN]')
-  }
-  if (capa & (1 << 29)) {
-    parsedOutput.push('[TX-ANTENNA-PATTERN]')
-  }
-
-  let vhtAdapt = (capa & 0xc000000) >> 26
-  if (vhtAdapt == 2 || vhtAdapt == 3) {
-    parsedOutput.push('[VHT-LINK-ADAPT' + vhtAdapt + ']')
-  }
-
-  return parsedOutput
-}
-
-const filterHTCapabilities = (capstr, defaultstr) => {
-  let available = mapHTCapabilities(capstr)
-  let settings = defaultstr.split('[')
-  for (let setting of settings) {
-    if (setting == '') continue
-    setting = '[' + setting
-    if (available.indexOf(setting) == -1) {
-      defaultstr = defaultstr.replace(setting, '')
-    }
-  }
-  return defaultstr
-}
-
-const filterVHTCapabilities = (capstr, defaultstr) => {
-  let available = mapVHTCapabilities(capstr)
-  let settings = defaultstr.split('[')
-  for (let setting of settings) {
-    if (setting == '') continue
-    setting = '[' + setting
-    if (available.indexOf(setting) == -1) {
-      defaultstr = defaultstr.replace(setting, '')
-    }
-  }
-  return defaultstr
-}
-
-const filterCapabilities = (template, ht_capstr, vht_capstr, band) => {
-  //deep copy template
-  let config = JSON.parse(JSON.stringify(template))
-  if (band == 1) {
-    //delete vht_capab from config
-    if ('vht_capab' in config) {
-      delete config.vht_capab
-    }
-    //filter ht
-    config.ht_capab = filterHTCapabilities(ht_capstr, config.ht_capab)
-  } else if (band == 2) {
-    //filter ht and vht capab in config
-    config.ht_capab = filterHTCapabilities(ht_capstr, config.ht_capab)
-    config.vht_capab = filterVHTCapabilities(vht_capstr, config.vht_capab)
-  } else if (band == 4) {
-    //6ghz: anythign else?
-    //filter both ht and vht
-    config.ht_capab = filterHTCapabilities(ht_capstr, config.ht_capab)
-    config.vht_capab = filterVHTCapabilities(vht_capstr, config.vht_capab)
-  }
-
-  return config
-}
-
 const WifiHostapd = (props) => {
   const context = useContext(AlertContext)
   const [iface, setIface] = useState('')
@@ -448,28 +119,6 @@ const WifiHostapd = (props) => {
   ]
   const canEdit = canEditInt.concat(canEditString)
 
-  const sortConf = (conf) => {
-    // put the ones we can change at the top
-    //force ssid at the top
-    return Object.keys(conf)
-      .sort((a, b) => {
-        if (a == 'ssid') {
-          return -1
-        }
-        if (canEdit.includes(a)) {
-          if (canEdit.includes(b)) {
-            return a > b ? 1 : a < b ? -1 : 0
-          }
-          return -1
-        }
-
-        return a > b ? 1 : a < b ? -1 : 0
-      })
-      .reduce((obj, key) => {
-        obj[key] = conf[key]
-        return obj
-      }, {})
-  }
 
   const updateCapabilitiesTooltips = () => {
     //update the tooltips for vht_capab, ht_capab
@@ -477,16 +126,16 @@ const WifiHostapd = (props) => {
     let ht_capab, vht_capab
     if (config.hw_mode == 'a') {
       //this assumes 5ghz, need to handle 6ghz and wifi 6
-      ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 2)
+      ;[ht_capab, vht_capab] = generateCapabilitiesString(iwMap, iface, 2)
     } else if (config.hw_mode == 'g' || config.hw_mode == 'b') {
-      ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 1)
+      ;[ht_capab, vht_capab] = generateCapabilitiesString(iwMap, iface, 1)
     } else {
-      ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 2)
+      ;[ht_capab, vht_capab] = generateCapabilitiesString(iwMap, iface, 2)
       if (ht_capab == '' && vht_capab == '') {
-        ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 1)
+        ;[ht_capab, vht_capab] = generateCapabilitiesString(iwMap, iface, 1)
       }
       if (ht_capab == '' && vht_capab == '') {
-        ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 4)
+        ;[ht_capab, vht_capab] = generateCapabilitiesString(iwMap, iface, 4)
       }
     }
 
@@ -496,33 +145,8 @@ const WifiHostapd = (props) => {
   }
 
   const updateCapabilities = () => {
-    let new_config = { ...config }
-
     //enable all capabilities
-
-    let ht_capab, vht_capab
-    if (config.hw_mode == 'a') {
-      //this assumes 5ghz, need to handle 6ghz and wifi 6
-      ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 2)
-    } else if (config.hw_mode == 'g' || config.hw_mode == 'b') {
-      ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 1)
-    } else {
-      ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 2)
-      if (ht_capab == '' && vht_capab == '') {
-        ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 1)
-      }
-      if (ht_capab == '' && vht_capab == '') {
-        ;[ht_capab, vht_capab] = generateCapabilitiesString(iface, 4)
-      }
-    }
-    if (ht_capab) {
-      ht_capab.sort()
-      new_config.ht_capab = ht_capab.join('')
-    }
-    if (vht_capab) {
-      vht_capab.sort()
-      new_config.vht_capab = vht_capab.join('')
-    }
+    let new_config = getBestWifiConfig(iwMap, iface, config)
 
     pushConfig(new_config)
   }
@@ -544,11 +168,17 @@ const WifiHostapd = (props) => {
           return iw;
         }));
 
+
         //make a phy to iws map and devname to iw map
         let iwMap = {}
         iws.forEach((iw) => {
           iwMap[iw.wiphy] = iw
           Object.keys(iw.devices).forEach((dev) => {
+            if (isSPRCompat(iw)) {
+              iw.spr_compat = true
+            } else {
+              iw.spr_compat = false
+            }
             iwMap[dev] = iw
             iwMap[iw.wiphy].dev = dev
           })
@@ -558,6 +188,7 @@ const WifiHostapd = (props) => {
           iw.failsafeStatus = await wifiAPI.checkFailsafe(iwMap[iw.wiphy].dev);
           return iw;
         }));
+
         setIwMap(iwMap)
         setIws(iws)
       })
@@ -603,6 +234,7 @@ const WifiHostapd = (props) => {
         setConfig(sortConf(conf))
       })
       .catch((err) => {
+        context.error('failed to retrieve configuration for ' + iface, err)
         //configuration not found. How to handle?
         setConfig({})
       })
@@ -671,87 +303,6 @@ const WifiHostapd = (props) => {
     commitConfig()
   }
 
-  const generateCapabilitiesString = (iface, wanted_band) => {
-    let iw_info = iwMap[iface]
-    let ht_capstr = ''
-    let vht_capstr = ''
-
-    if (iw_info == undefined) {
-      return ['', '']
-    }
-
-    for (let i = 0; i < iw_info.bands.length; i++) {
-      let band = iw_info.bands[i].band
-      let ht_capstr = ''
-      let vht_capstr = ''
-
-      if (iw_info.bands[i].capabilities != undefined) {
-        ht_capstr = iw_info.bands[i].capabilities[0]
-      }
-
-      if (iw_info.bands[i].vht_capabilities != undefined) {
-        vht_capstr = iw_info.bands[i].vht_capabilities[0]
-      }
-
-      if (band.includes('Band 2') && wanted_band == 2) {
-        return [mapHTCapabilities(ht_capstr), mapVHTCapabilities(vht_capstr)]
-      } else if (band.includes('Band 1') && wanted_band == 1) {
-        return [mapHTCapabilities(ht_capstr), mapVHTCapabilities(vht_capstr)]
-      } else if (band.includes('Band 4') && wanted_band == 4) {
-        return [mapHTCapabilities(ht_capstr), mapVHTCapabilities(vht_capstr)]
-      }
-    }
-
-    return ['', '']
-  }
-
-  const generateConfigForBand = (iface, wanted_band) => {
-    let iw_info = iwMap[iface]
-    let ht_capstr = ''
-    let vht_capstr = ''
-
-    let defaultConfig
-
-    for (let i = 0; i < iw_info.bands.length; i++) {
-      let band = iw_info.bands[i].band
-      let ht_capstr = ''
-      let vht_capstr = ''
-
-      if (iw_info.bands[i].capabilities != undefined) {
-        ht_capstr = iw_info.bands[i].capabilities[0]
-      }
-
-      if (iw_info.bands[i].vht_capabilities != undefined) {
-        vht_capstr = iw_info.bands[i].vht_capabilities[0]
-      }
-
-      if (band.includes('Band 2') && wanted_band == 2) {
-        defaultConfig = filterCapabilities(
-          default5Ghz,
-          ht_capstr,
-          vht_capstr,
-          2
-        )
-        break
-      } else if (band.includes('Band 1') && wanted_band == 1) {
-        defaultConfig = filterCapabilities(
-          default2Ghz,
-          ht_capstr,
-          vht_capstr,
-          1
-        )
-      } else if (band.includes('Band 4') && wanted_band == 4) {
-        defaultConfig = filterCapabilities(
-          default6GHz,
-          ht_capstr,
-          vht_capstr,
-          4
-        )
-      }
-    }
-
-    return defaultConfig
-  }
 
   const generateHostAPConfiguration = () => {
     //band 1 -> 2.4GHz
@@ -771,37 +322,15 @@ const WifiHostapd = (props) => {
     //iterate through bands, looking for band 2, band 1, band 4 in that order
 
     defaultConfig =
-      generateConfigForBand(iface, 2) ||
-      generateConfigForBand(iface, 1) ||
-      generateConfigForBand(iface, 4)
+      generateConfigForBand(iwMap, iface, 2) ||
+      generateConfigForBand(iwMap, iface, 1) ||
+      generateConfigForBand(iwMap, iface, 4)
 
     if (defaultConfig == undefined) {
       context.error(
         'could not determine default hostap configuration. using a likely broken default'
       )
       return
-    }
-
-    let has_wifi6 = false
-
-    for (let i = 0; i < iw_info.bands.length; i++) {
-      let band = iw_info.bands[i]
-      if (band.he_phy_capabilities) {
-        has_wifi6 = true
-        break
-      }
-    }
-
-    if (has_wifi6 == true) {
-      defaultConfig.he_mu_beamformer = '1'
-      defaultConfig.he_su_beamformee = '1'
-      defaultConfig.he_su_beamformer = '1'
-      defaultConfig.ieee80211ax = '1'
-    } else {
-      defaultConfig.he_mu_beamformer = '0'
-      defaultConfig.he_su_beamformee = '0'
-      defaultConfig.he_su_beamformer = '0'
-      defaultConfig.ieee80211ax = '0'
     }
 
     pushConfig(defaultConfig)
@@ -855,7 +384,7 @@ const WifiHostapd = (props) => {
         (config.vht_capab == undefined || config.vht_capab == '')
       ) {
         //re-enable vht capabilities for 5GHz
-        let tempConfig = generateConfigForBand(iface, 2)
+        let tempConfig = generateConfigForBand(iwMap, iface, 2)
         data.Vht_capab = tempConfig.vht_capab
       }
 
@@ -916,7 +445,7 @@ const WifiHostapd = (props) => {
       devsSelect.push({
         label,
         value: _iface,
-        isDisabled: _iface.includes('.')
+        isDisabled: _iface.includes('.') || (iwMap[_iface].spr_compat == false)
       })
 
       if (_iface == iface) {
